MFC 286370: Add various tests to ensure that invalid arguments passed to mmap() trigger failures.
Note: most of the tests that should provoke an EINVAL error do not
pass on stable/10 as stable/10 does not have the changes that added
more strict parameter checking to mmap().  Those changes will not be
merged to stable/10, so I have disabled them via #if 0 with a comment
explaining why.


git-svn-id: https://svn.freebsd.org/base/stable/10@288466 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f