MFC r281785
Always send log(9) messages to the message buffer.

It is truer to the semantics of logging for messages to *always*
go to the message buffer, where they can eventually be collected
and, in fact, be put into a log file.

This restores the behavior prior to r70239, which seems to have
changed it inadvertently.

Submitted by:   Eric Badger <eric@badgerio.us>
Obtained from:  Dell Inc.


git-svn-id: https://svn.freebsd.org/base/stable/10@288497 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f