It appears that under some circumstances, like virtualisiation, the 'rdrand' instruction may occasionally not return random numbers, in spite of looping attempts to do so. The reusult is a KASSERT/panic.
Reluctantly accept this state-of-affairs, but make a noise about it.
if this 'noise' spams the console, it may be time to discontinue
using that source.

This is written in a general way to account for /any/ source that
might not supply random numbers when required.

Submitted by:	jkh (report and slightly different fix)
Approved by:	so (/dev/random blanket)


git-svn-id: https://svn.freebsd.org/base/head@288703 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f