If world is built with a custom sendmail.cf, use it for the distribution target.  This is the feeder for mergemaster / etcupdate.  This change makes installworld/mergemaster/etcupdate behave the same regardless of whether SENDMAIL_MC or SENDMAIL_CF is used.
If you use a custom SENDMAIL_MC/CF in make.conf and excluded it from
mergemaster.rc/etcupdate.conf to work around the conflicts, you may wish
to revert that or change it from 'ignore' to 'always install'.

If you do not use a custom SENDMAIL_MC/CF, there should be no change in
behavior.


git-svn-id: https://svn.freebsd.org/base/head@289200 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f