Switch mips busdma to using the common busdma_buffalloc code.  This amounts to copying in some code from the armv4 busdma, and adapting a few variable and flag names to match the surrounding mips code.
Instead of keeping a local cache of prealloced busdma_map structs on a
mutex-protected list, set up an uma zone to cache them.

Instead of all memory allocations using M_DEVBUF, use new categories
M_BUSDMA for allocations of metadata (tags, maps, segment tracking lists),
and M_BOUNCE for bounce pages.

When buffers are allocated out of the busdma_bufalloc zones the alignment
and size of the buffers is known, and the code can skip doing any "partial
cacheline flush" logic to preserve data that may be adjacent to the DMA
buffer but contain non-DMA data.

Reviewed by:	adrian, imp


git-svn-id: https://svn.freebsd.org/base/head@289701 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f